+++ build.gradle.kts
@@ -1,10 +1,9 @@
 plugins {
-    kotlin("jvm") version libs.versions.kotlin.get()
+    kotlin("multiplatform") version libs.versions.kotlin.get()
     kotlin("plugin.serialization") version libs.versions.kotlin.get()
     `maven-publish`
     id("app.cash.sqldelight") version "2.0.1"
 }
-
 
 fun getBranchName(): String {
     val branchNameEnv = System.getenv("BRANCH_NAME")
@@ -35,107 +34,98 @@ repositories {
     mavenCentral()
 }
 
-dependencies {
-    // Arrow dependencies from version catalog
-    implementation(libs.arrow.core)
-    implementation(libs.arrow.fx.coroutines)
-
-    // Additional dependencies
-    implementation(libs.ulid.creator)
-    implementation(libs.kotlinx.coroutines.core)
-    implementation(libs.kotlinx.serialization.json)
-    implementation(libs.ktor.client.core)
-    implementation(libs.ktor.client.cio)
-    implementation(libs.ktor.client.content.negotiation)
-    implementation(libs.ktor.serialization.kotlinx.json)
-
-    // SQLDelight dependencies
-    implementation("app.cash.sqldelight:sqlite-driver:2.0.1")
-    implementation("app.cash.sqldelight:coroutines-extensions:2.0.1")
-
-    // Test dependencies
-    testImplementation(libs.kotest.runner.junit5)
-    testImplementation(libs.kotest.assertions.core)
-    testImplementation(libs.kotlinx.coroutines.test)
-    testImplementation("io.kotest:kotest-runner-junit5:5.9.1")
-    testImplementation("io.kotest:kotest-assertions-core:5.8.0")
-    testImplementation("io.kotest:kotest-property:5.9.1")
-
-    testImplementation("io.ktor:ktor-client-content-negotiation:3.1.1")
-    testImplementation("io.ktor:ktor-serialization-kotlinx-json:3.1.1")
-}
-
-tasks.withType<Test>().configureEach {
-    useJUnitPlatform()
-}
-
 kotlin {
-    jvmToolchain(19)
+    // Define the targets you want to support
+    jvm() // Configures JVM source sets (jvmMain, jvmTest)
+
+    // Define desktop targets and group them
+    // Use presets if you prefer, but explicit definition is clear
+    linuxX64()
+    macosX64()
+    mingwX64() // Windows
+
+    sourceSets {
+
+        val commonMain by getting {
+            dependencies {
+                // Dependencies that work on all platforms go here
+                implementation(libs.arrow.core)
+                implementation(libs.arrow.fx.coroutines)
+
+                // Check if libs.ulid.creator supports multiplatform or find an alternative
+                implementation(libs.ulid.kotlin)
+
+                // kotlinx.coroutines and serialization core are multiplatform
+                implementation(libs.kotlinx.coroutines.core)
+                implementation(libs.kotlinx.serialization.json)
+
+                // Ktor client core is multiplatform
+                implementation(libs.ktor.client.core)
+                implementation(libs.ktor.client.content.negotiation)
+                implementation(libs.ktor.serialization.kotlinx.json)
+
+                // SQLDelight runtime for multiplatform
+                implementation("app.cash.sqldelight:runtime:2.0.1")
+                // Coroutines extensions might be multiplatform or need platform specifics
+                implementation("app.cash.sqldelight:coroutines-extensions:2.0.1")
+
+            }
+        }
+
+        val commonTest by getting {
+            dependencies {
+
+            }
+        }
+
+        val jvmMain by getting {
+            dependencies {
+                // JVM-specific dependencies
+                implementation("app.cash.sqldelight:sqlite-driver:2.0.1") // JVM SQLite driver
+                implementation(libs.ktor.client.cio) // JVM Ktor engine
+            }
+        }
+
+        val jvmTest by getting {
+            dependencies {
+                // JVM-specific test dependencies if any
+                // Add the JUnit 5 runner for JVM tests
+                implementation(libs.kotest.runner.junit5)
+            }
+        }
+
+        val desktopMain by creating {
+            dependsOn(commonMain) // Inherit dependencies from commonMain
+            dependencies {
+                implementation("io.ktor:ktor-client-curl:3.1.1")
+
+            }
+        }
+
+        val desktopTest by creating {
+            dependencies {
+                implementation("io.kotest:kotest-runner-native:5.9.1")
+            }
+        }
+        linuxX64Main.get().dependsOn(desktopMain)
+        mingwX64Main.get().dependsOn(desktopMain)
+        macosX64Main.get().dependsOn(desktopMain)
+    }
+}
+
+// Configure test tasks. JVM tests use JUnit Platform.
+// The KMP plugin handles native test tasks.
+tasks.withType<Test>().configureEach {
+    // This configuration typically applies only to the JVM test task
+    useJUnitPlatform()
 }
 
 // --- Publishing Configuration ---
 publishing {
-    publications {
-        // Create a publication named 'maven'. You can name it anything.
-        create<MavenPublication>("maven") {
-            // Apply the component generated by the kotlin("jvm") plugin.
-            from(components["java"])
-
-            // Set the artifact details (group and version are inherited)
-            // artifactId defaults to project.name (your directory name), override if needed
-            // artifactId = "my-library-artifact"
-
-            // --- POM Metadata (Customize this!) ---
-            pom {
-                properties.set(mapOf("branchName" to getBranchName()))
-                name.set("KAIA")
-                description.set("A concise description of your library.")
-                url.set("https://github.com/GabrielOlvH/kaia")
-                licenses {
-                    license {
-                        name.set("MIT License")
-                        url.set("https://opensource.org/licenses/MIT")
-                    }
-                }
-                developers {
-                    developer {
-                        id.set("gabrielolvh")
-                        name.set("Gabriel Oliveira")
-                        email.set("gabrielh.oliveira222@gmail.com")
-                    }
-                }
-                scm {
-                    connection.set("scm:git:git://github.com/dev-gabrielolvh/kaia.git")
-                    developerConnection.set("scm:git:ssh://github.com:gabrielolvh/kaia.git")
-                    url.set("https://github.com/gabrielolvh/kaia/tree/main")
-                }
-            }
-        }
-    }
-    repositories {
-        // Configure Sonatype Nexus repository
-        maven {
-            name = "SonatypeNexus"
-            
-            // Dynamically determine repository URL based on version (SNAPSHOT vs Release)
-            url = uri(
-                if (version.toString().endsWith("SNAPSHOT")) {
-                    "https://maven.gabrielolv.dev/repository/maven-snapshots/"
-                } else {
-                    "https://maven.gabrielolv.dev/repository/maven-releases/"
-                }
-            )
-            
-            credentials {
-                // Use environment variables or gradle.properties for credentials
-                username = System.getenv("NEXUS_USERNAME") ?: project.findProperty("nexus.username") as String?
-                password = System.getenv("NEXUS_PASSWORD") ?: project.findProperty("nexus.password") as String?
-            }
-        }
-    }
 }
 
 // Optional: Ensure publish task runs after build
 tasks.withType<PublishToMavenRepository>().configureEach {
-    dependsOn(tasks.build)
+    dependsOn(tasks.build) // Consider if 'build' is appropriate for multiplatform publish
+    // You might need to depend on specific tasks like `publishKotlinPublicationToSonatypeNexusRepository`
 }
