+++ src/commonMain/kotlin/dev/gabrielolv/kaia/core/tools/ToolExecutionFailedException.kt
