+++ src/commonMain/kotlin/dev/gabrielolv/kaia/core/tools/ToolResult.kt
