+++ src/commonMain/kotlin/dev/gabrielolv/kaia/core/tools/typed/TypedTool.kt
