+++ src/commonMain/kotlin/dev/gabrielolv/kaia/core/tools/typed/validation/ValidationRule.kt
