+++ src/commonMain/kotlin/dev/gabrielolv/kaia/core/tools/typed/validation/ValidationRules.kt
