+++ src/commonMain/kotlin/dev/gabrielolv/kaia/utils/HttpClient.kt
@@ -1,0 +1,5 @@
+package dev.gabrielolv.kaia.utils
+
+import io.ktor.client.*
+
+expect val httpClient: HttpClient
