+++ src/desktopMain/kotlin/dev/gabrielolv/kaia/utils/HttpClient.kt
@@ -1,0 +1,23 @@
+package dev.gabrielolv.kaia.utils
+
+import io.ktor.client.*
+import io.ktor.client.engine.curl.*
+import io.ktor.client.plugins.contentnegotiation.*
+import io.ktor.serialization.kotlinx.json.*
+import kotlinx.serialization.ExperimentalSerializationApi
+import kotlinx.serialization.json.Json
+import kotlinx.serialization.json.JsonNamingStrategy
+
+@OptIn(ExperimentalSerializationApi::class)
+actual val httpClient: HttpClient = HttpClient(Curl) { // Use the Curl engine
+    install(ContentNegotiation) {
+        json(Json {
+            ignoreUnknownKeys = true
+            explicitNulls = false
+            encodeDefaults = true
+            isLenient = true
+            namingStrategy = JsonNamingStrategy.SnakeCase
+        })
+    }
+    expectSuccess = true
+}
